requirements.txt
accelerate==1.12.0
datasets==4.4.1
dill==0.4.0
huggingface_hub==1.2.3
numpy==2.3.5
peft==0.18.0
pillow==12.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
qwen-vl-utils==0.0.14
safetensors==0.7.0
timm==1.0.22
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
transformers==5.0.0rc1
trl==0.26.2
typing_extensions==4.15.0
zai==0.0.2
